testing/web-platform/tests/XMLHttpRequest/setrequestheader-bogus-name.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: setRequestHeader() name argument checks</title>
    <meta charset="utf-8">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-setrequestheader()-method" data-tested-assertations="/following::ol/li[3]" />  
  </head>
  <body>
    <div id="log"></div>
    <script>
      function try_name(name) {
        test(function() {
          var client = new XMLHttpRequest()
          client.open("GET", "...")
          assert_throws("SyntaxError", function() { client.setRequestHeader(name, 'x-value') })
        })
      }
      try_name("")
      try_name("t\rt")
      try_name("t\nt")
      try_name("t: t")
      try_name("t:t")
      try_name("t<t")
      try_name("t t")
      try_name(" tt")
      try_name(":tt")
      try_name("\ttt")
      try_name("\vtt")
      try_name("t\0t")
      try_name("t\"t")
      try_name("t,t")
      try_name("t;t")
      try_name("()[]{}")
      try_name("a?B")
      try_name("X-ﾃｽﾄ")
      try_name("a=B")
    </script>
  </body>
</html>
